Woke up really early and couldn't go back to sleep.
Someone on #sopel was asking how to list channels, so I decided to write
this in case they got stuck figuring out how. In the process, I got
stuck myself a few times, but it was worth learning more about how to do
things with IRC lines that aren't PRIVMSGs.